requirements.txt
Django>=4.2.0,<5.0.0
mysqlclient>=2.2.0


